requirements.txt
flask==3.1.1
flask-cors==6.0.1
transformers==4.52.4
torch==2.7.1
accelerate==1.8.1
python-dotenv==1.1.0
markdown==3.8.2
pygments==2.19.1
numpy==2.3.0
requests==2.32.4
huggingface-hub==0.33.0
tokenizers==0.21.1
safetensors==0.5.3
tqdm==4.67.1 